Removed wiremod support to fix #1
This fixes #1 but removes Wiremod functionality for now.

Since the addon used Wiremod in a very wierd fashion the addon was
basically overriding everything Wiremod did to entities. This is now
solved by just removing (most of) the wiremod functionality and
associated files (wm_cl_init.lua was very nasty).

I'll look into creating a version that works with Wiremod, but it's
gonna take some research.